requirements/base.txt
Django==4.1.3
python-dotenv==1.0.0
kubernetes==28.1.0
kafka-python==2.0.2
confluent_kafka==2.3.0